requirements.txt
protobuf
pycryptodome
pyyaml